move utils file, disable validation for now
